Fetch film details from API rather than db
Since there is no limit on API I can fetch all the details from the API
rather than storing them in the database. This will also help in keeping
the data up to date.

Added a presenter to pass the film details to the view. Added new API
call that appends credits (directors) to the view.